refactor(app): split file tool tests
Move the large inline tool-file test module out of tools/file.rs into tools/file/tests.rs.

Keep the move mechanical and preserve the existing tool-file cfg on the parent module declaration.
